+++ sampletv/src/main/java/testlibuiza/activity/PlayerCustomActivity.java
@@ -51,10 +51,10 @@ public class PlayerCustomActivity extends BaseActivity implements UZCallback, UZ
         rl = (RelativeLayout) findViewById(R.id.rl);
         recyclerView = (RecyclerView) findViewById(R.id.recycler_view);
         uzVideo = (UZVideo) findViewById(R.id.uiza_video);
-        uzVideo.setUZCallback(this);
-        uzVideo.setUZTVCallback(this);
+        uzVideo.addUZCallback(this);
+        uzVideo.addUZTVCallback(this);
         uzVideo.setDefaultUseController(false);
-        uzVideo.setOnTouchEvent(new UZPlayerView.OnTouchEvent() {
+        uzVideo.addOnTouchEvent(new UZPlayerView.OnTouchEvent() {
             @Override
             public void onSingleTapConfirmed(float x, float y) {
             }